Moving test_noisy into test_aer_executor
